+++ src/tap/communication/serial/ref_serial.hpp
@@ -181,10 +181,9 @@ private:
      */
     bool decodeToSiteEventData(const ReceivedSerialMessage& message);
     /**
-     * Decodes ref serial message containing warning information (if a robot on your team received a
-     * yellow or red card).
+     * REF_MESSAGE_TYPE_WARNING_DATA
      */
-    bool decodeToWarningData(const ReceivedSerialMessage& message);
+    bool decodeToWarningData(const ReceivedSerialMessage &message);
     /**
      * Decodes ref serial message containing the firing/driving heat limits and cooling
      * rates for the robot.