+++ src/tap/communication/serial/ref_serial_data.hpp
@@ -253,14 +253,6 @@ public:
                                                           ///< message was received
         };
 
-        /**
-         * Ref serial warning data. This informational struct is intended to be updated each time a
-         * yellow or red card has been received.
-         *
-         * When a yellow card is received, the offending Operator will be blocked from seeing
-         * anything for 5 seconds, while the operation interfaces of other operators in the
-         * offending team will be blocked for 2 seconds.
-         */
         struct RefereeWarningData
         {
             uint8_t level;        /**<
@@ -269,12 +261,9 @@ public:
                                    * 2: Red card
                                    * 3: Forfeiture
                                    */
-            RobotId foulRobotID;  ///< The robot that received the referee warning
+            uint8_t foulRobotID;  ///< The robot that received the referee warning
             uint32_t lastReceivedWarningRobotTime = 0;  ///< Last time (in milliseconds) that a
                                                         ///< warning was received.
-
-            static constexpr uint32_t OFFENDING_OPERATOR_BLIND_TIME = 5'000;
-            static constexpr uint32_t NONOFFENDING_OPERATOR_BLIND_TIME = 2'000;
         };
 
         struct RobotData